feat!: allow using CDN with token (for queries)
BREAKING CHANGE: when specifying both `useCdn: true` and a `token`, the client
previously warned and did not use the CDN for queries, instead opting for the
"live"/"uncached" API.

With this change, using the same configuration will now use the CDN API, which
means you might get cached (potentially stale) responses, but is more in line
with what the configuration is telling it to do.

This is possible because of a change to the API CDN that allows it to include
the authorization token as part of the cache key, ensuring that two different
tokens (or a request with and a request without a token) are treated as
different cache entities.