fix: undeprecate `request()` (#205)
The reason it was marked as deprecated was to steer users away from it in favor
or using their own request library, since we didn't want the maintenance burden
of having to keep request options/behavior compatible between releases. However,
after seeing a number of real-world use cases where users want to reuse the
client to do requests against Sanity API endpoints that do not have their own
built-in method (scheduled publishing, webhook APIs etc), and the complexity of
figuring out the authentication part (cookie? auth token?), we decided it is a
reasonable ask to allow using the method.